Extract superclass HealthApp
… and dissolve CommonEndpointSpecs
